ut_webGIS/webGIS_Django_for _interactions_carshops/car/carlink/requirements.txt
django==5.2.1
django-allauth
django-crispy-forms
crispy-bootstrap5
psycopg2-binary  # برای PostGIS